Refactoring, bugfixes
