+++ tests/tibanna/unicorn/test_ec2_utils.py
@@ -4,8 +4,10 @@ from tibanna.ec2_utils import (
     Config,
     Execution,
     upload_workflow_to_s3,
-    get_file_size,
-    cost_estimate
+    get_file_size
+)
+from tibanna.pricing_utils import (
+    get_cost_estimate
 )
 from tibanna.utils import create_jobid
 from tibanna.exceptions import (
@@ -778,7 +780,7 @@ def test_ec2_cost_estimate_missing_availablity_zone():
 
     postrunjsonobj = json.loads(postrunjsonstr)
     postrunjson = AwsemPostRunJson(**postrunjsonobj)
-    assert cost_estimate(postrunjson) == 0.0
+    assert get_cost_estimate(postrunjson) == 0.0
 
 
 def test_ec2_cost_estimate_medium_nonspot():
@@ -793,7 +795,7 @@ def test_ec2_cost_estimate_medium_nonspot():
     aws_price_overwrite = {
         'ec2_ondemand_price': 0.0416, 'ebs_root_storage_price': 0.08
         } 
-    estimate = cost_estimate(postrunjson, aws_price_overwrite=aws_price_overwrite)
+    estimate = get_cost_estimate(postrunjson, aws_price_overwrite=aws_price_overwrite)
     assert estimate == 0.004384172839506173
 
 
@@ -809,7 +811,7 @@ def test_ec2_cost_estimate_small_spot():
     aws_price_overwrite = {
         'ec2_spot_price': 0.0064, 'ebs_root_storage_price': 0.08
         } 
-    estimate = cost_estimate(postrunjson, aws_price_overwrite=aws_price_overwrite)
+    estimate = get_cost_estimate(postrunjson, aws_price_overwrite=aws_price_overwrite)
     assert estimate == 0.0009326172839506175
 
 
@@ -825,7 +827,7 @@ def test_ec2_cost_estimate_small_spot_gp3_iops():
     aws_price_overwrite = {
         'ec2_spot_price': 0.0064, 'ebs_root_storage_price': 0.08, 'ebs_gp3_iops_price': 0.005
         } 
-    estimate = cost_estimate(postrunjson, aws_price_overwrite=aws_price_overwrite)
+    estimate = get_cost_estimate(postrunjson, aws_price_overwrite=aws_price_overwrite)
     assert estimate == 0.0012730879629629633
 
 
@@ -841,6 +843,6 @@ def test_ec2_cost_estimate_small_spot_io2_iops():
     aws_price_overwrite = {
         'ec2_spot_price': 0.0064, 'ebs_root_storage_price': 0.08, 'ebs_storage_price': 0.125, 'ebs_io2_iops_prices': [0.065, 0.0455, 0.03185]
         } 
-    estimate = cost_estimate(postrunjson, aws_price_overwrite=aws_price_overwrite)
+    estimate = get_cost_estimate(postrunjson, aws_price_overwrite=aws_price_overwrite)
     assert estimate == 0.029224481481481476
