+++ tibanna/core.py
@@ -45,8 +45,13 @@ from .utils import (
 )
 from .ec2_utils import (
     UnicornInput,
-    upload_workflow_to_s3,
-    cost_estimate
+    upload_workflow_to_s3
+)
+from .pricing_utils import (
+    get_cost,
+    get_cost_estimate,
+    update_cost_estimate_in_tsv,
+    update_cost_in_tsv
 )
 from .ami import AMI
 # from botocore.errorfactory import ExecutionAlreadyExists
@@ -1057,47 +1062,29 @@ class API(object):
             webbrowser.open(METRICS_URL(log_bucket, job_id))
 
     def cost_estimate(self, job_id, update_tsv=False):
-
-        # We return the real cost, if it is availble, but don't automatically update the Cost row in the tsv
-        precise_cost = self.cost(job_id, update_tsv=False)
-        if(precise_cost and precise_cost > 0.0):
-            return precise_cost
-
         postrunjsonstr = self.log(job_id=job_id, postrunjson=True)
         if not postrunjsonstr:
             logger.info("Cost estimation error: postrunjson not found")
             return 0.0
         postrunjsonobj = json.loads(postrunjsonstr)
         postrunjson = AwsemPostRunJson(**postrunjsonobj)
-        
+        log_bucket = postrunjson.config.log_bucket
+
+        # We return the real cost, if it is availble, but don't automatically update the Cost row in the tsv
+        precise_cost = self.cost(job_id, update_tsv=False)
+        if(precise_cost and precise_cost > 0.0):
+            if update_tsv:
+                update_cost_estimate_in_tsv(log_bucket, job_id, precise_cost)
+            return precise_cost
+
         # awsf_image was added in 1.0.0. We use that to get the correct ebs root type
         ebs_root_type = 'gp3' if 'awsf_image' in postrunjsonobj['config'] else 'gp2'
 
-        cost = cost_estimate(postrunjson, ebs_root_type)
+        cost = get_cost_estimate(postrunjson, ebs_root_type)
 
         if update_tsv:
-            log_bucket = postrunjson.config.log_bucket
-            # reading from metrics_report.tsv
-            does_key_exist(log_bucket, job_id + '.metrics/metrics_report.tsv')
-            read_file = read_s3(log_bucket, os.path.join(job_id + '.metrics/', 'metrics_report.tsv'))
-
-            write_file = ""
-            for row in read_file.splitlines():
-                # Remove Estimated_Cost from file, since we want to update it
-                if("Estimated_Cost" not in row.split("\t")):
-                    write_file = write_file + row + '\n'
-
-            write_file = write_file + 'Estimated_Cost\t' + str(cost) + '\n'
-
-            # writing
-            with open('metrics_report.tsv', 'w') as fo:
-                fo.write(write_file)
-            # upload new metrics_report.tsv
-            upload('metrics_report.tsv', log_bucket, job_id + '.metrics/')
-            os.remove('metrics_report.tsv')
-
+            update_cost_estimate_in_tsv(log_bucket, job_id, cost)
         return cost
-
 
     def cost(self, job_id, sfn=None, update_tsv=False):
         if not sfn:
@@ -1106,50 +1093,15 @@ class API(object):
         if not postrunjsonstr:
             return None
         postrunjson = AwsemPostRunJson(**json.loads(postrunjsonstr))
-        job = postrunjson.Job
 
-        def reformat_time(t, delta):
-            d = datetime.strptime(t, '%Y%m%d-%H:%M:%S-UTC') + timedelta(days=delta)
-            return d.strftime("%Y-%m-%d")
+        cost = get_cost(postrunjson, job_id)
 
-        start_time = reformat_time(job.start_time, -1)  # give more room
-        end_time = reformat_time(job.end_time, 1)  # give more room
-        billing_args = {'Filter': {'Tags': {'Key': 'Name', 'Values': ['awsem-' + job_id]}},
-                        'Granularity': 'DAILY',
-                        'TimePeriod': {'Start': start_time,
-                                       'End': end_time},
-                        'Metrics': ['BlendedCost'],
-                        }
-        
-        try:
-            billingres = boto3.client('ce').get_cost_and_usage(**billing_args)
-        except botocore.exceptions.ClientError as e:
-            logger.warning("%s. Please try to deploy the latest version of Tibanna." % e)
-            return 0.0
-
-        cost = sum([float(_['Total']['BlendedCost']['Amount']) for _ in billingres['ResultsByTime']])
         if update_tsv:
             log_bucket = postrunjson.config.log_bucket
-            # reading from metrics_report.tsv
-            does_key_exist(log_bucket, job_id + '.metrics/metrics_report.tsv')
-            read_file = read_s3(log_bucket, os.path.join(job_id + '.metrics/', 'metrics_report.tsv'))
-
-            write_file = ""
-            for row in read_file.splitlines():
-                # Remove Cost from file, since we want to update it
-                if("Cost" not in row.split("\t")):
-                    write_file = write_file + row + '\n'
-
-            write_file = write_file + 'Cost\t' + str(cost) + '\n'
-
-            #writing
-            with open('metrics_report.tsv', 'w') as fo:
-                fo.write(write_file)
-            # upload new metrics_report.tsv
-            upload('metrics_report.tsv', log_bucket, job_id + '.metrics/')
-            os.remove('metrics_report.tsv')
+            update_cost_in_tsv(log_bucket, job_id, cost)
 
         return cost
+
 
     def does_dynamo_table_exist(self, tablename):
         try:
